Make the definition of the pullback more precise
B_y(ybar) had extra embedded information that isn't in the definition of the function. Higher order AD concerns can ask things like "what is the derivative of B_y(ybar) with respect to x?" which seemed to confuse some 18.337 students since x is left off in that notation. The notation B_f^x(ybar) on y=f(x) makes it explicit that there is an x-dependence in the pullback function, and this corresponds to how information about the forward pass is embedded in the closures that Zygote is generating. This notation tweak is quite small but seemed to make it a lot more clear, at least to me and for the way I was teaching it, so I thought I'd upstream it back to Zygote.