remove Flux page
Now that it's the default, this is well-covered by the Flux.jl documentation. The examples here don't really add much over the rest of the Zygote documentation (since such usage of gradient has previously been shown, sans the `params` piece) and so it seems out of place in what's generally a fairly concise documentation.