Fix GPU broadcasting tests
Tests for the correctness of gradients when doing basic broadcasting
over arrays on the GPU were added in #373. However, that was before
GPU-enabled CI was set up on the repository, so the tests as written
were not put through CI.

The issue with the tests as written are just that they use integers.
This commit changes them to `Float32`, which should fix the
`MethodError`s that occur when running these tests.