Documentation edit on implicit parameters
After struggling to handle implicit parameters with a Flux model, and following a discourse discussion (https://discourse.julialang.org/t/unrecognized-gradient-using-zygote-for-ad-with-universal-differential-equations/63791/2) , I have decided to add some extra details on how to access and when to use implicit parameters. I hope this helps new users like me to avoid wasting time looking for this.